[Protocol] Extract client into separate file and add some docs
