+++ src/protocols/quake/mod.rs
@@ -1,6 +1,10 @@
 
-pub mod types;
-
 pub mod one;
 pub mod two;
 pub mod three;
+
+/// All types used by the implementation.
+pub mod types;
+pub use types::*;
+
+mod client;
